requirements.txt
# 跨平台依赖
Pillow>=10.0.0
customtkinter>=5.2.0

# Windows 依赖
pywin32>=306; platform_system=="Windows"
pystray>=0.19.0; platform_system=="Windows"

# macOS 依赖
rumps>=0.4.0; platform_system=="Darwin"
pyobjc-framework-Cocoa>=9.0; platform_system=="Darwin"

# Linux 依赖  
pystray>=0.19.0; platform_system=="Linux"